[dart2wasm] Runtime generic function instantiation
Adds a slot to the vtable of all generic closures with a function
taking a generic closure and type arguments and returning a closure
representing the instantiation of the generic closure with those type
arguments. Instantiation operations then call this instantiation
function, fetched from the vtable of the closure to be instantiated.

The context of an instantiation closure contains the original closure
and the type arguments. The vtable of an instantiation closure
contains trampoline functions that call the corresponding entry point
in the original closure with the instantiated type arguments.

The instantiation functions are shared between all closures with the
same representation. The trampolines are shared across representations
for the same vtable entries.

For now, the instantiation closure just inherits the runtime type from
the original closure, which means that its runtime type will be
incorrect. When we support generic function types with runtime type
substitution, we can perform such substitution when instantiating
a closure.

Change-Id: I5d3a4d623c0673f9c2188f8a7ddd5b28b9404ac4
Cq-Include-Trybots: luci.dart.try:dart2wasm-linux-x64-d8-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/262201
Reviewed-by: Joshua Litt <joshualitt@google.com>
Commit-Queue: Aske Simon Christensen <askesc@google.com>