Migration: account for generic bounds when determining whether null checks are necessary.
Previously, if a variable with a generic type was null-checked, and
the migration engine did not add a question mark to the type of the
variable itself, e.g.:

    f<T extends Object?>(T x) {
      if (x == null) { ... }
    }

Then the migration tool would erroneously consider the null check
unnecessary.  It failed to account for the fact that the bound on the
generic type might be nullable.

This change fixes the tool so that it now properly accounts for the
bound on the generic type variable.

Bug: https://b.corp.google.com/issues/247496662
Change-Id: I2b131e75b7e3fc5db0b668f1064340d404442c82
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/264501
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>